Use RACK_ENV instead of SCRIPT_ENV to detect production mode
RACK_ENV is automatically set on Heroku to "production" and it's a
common method of switching environments.

This change makes commands easier to run as the developer doesn't have
to manually set the SCRIPT_ENV in most circumstances. It will also allow
us to set up other environments (like "test") quite easily.

Fixes #4.