Allow the whole database to be exported as CSV over the web
This caters for the most common scenario - a Data Scientist 👩‍🔬 wishes
to use their Big Data Analysis software (Excel 📊) to take a look at the
data. The easiest thing for them is to just get all the data we have and
not have to think about if it's been filtered for just a subsection of
the data.

As our database grows this may become impractical but let's worry about
that when it happens.