Add tasks to help check changes we're making to the database
This commit contains two Rake tasks to help us verify changes we're
making to the database don't affect the data.

`statistics:temp:pull_request_40:ruby` can be run before and after the
migration added in 5624b43 to check that the averages for the pollution
statistics are the same.

`statistics:temp:pull_request_40:sql` can only be run after the
migration. It does the same calculation but using the database's
built-in average calculation.

The output from all those tasks should be the same before and after the
migration (except, of course, you can't run the second task until after
the migration). If it's not then it means something in the data has
changed.