Don't accommodate unseen time formats
This method currently works when the datetime string being parsed does
not have a time zone on the end. I've never seen a time formatted like
that in this project, so I don't think it's worth accommodating for that
case at this point. We're storing the original string as well, so we can
re-run if we start seeing more time formats that this doesn't
understand.