Inline migration code
This removes this migration's dependence on outside code by inlining the
conversion code from the (already removed) Aqm::Reading class.

We're doing this as it allows this migration to stand alone and not be
affectec by changes to outside code (as has already happened by this
whole class being removed).