Add failing test for scraping the JSON
This sets up a VCR fixture for the JSON version of the recordings, and
then a test to confirm that the responses are turned into our expected
AqmRecords.

I imagine as this gets closer to being validated, we'll need something
like timecop for getting the times right.

See https://github.com/equivalentideas/westconnex_M4_East_Air_Quality_Monitoring/issues/29