Use AqmReading class to build data from scraped web page
Instead of extracting the data from the page within the scraper we're
switching to using our new AqmReading class. The idea is that we should
then be able to switch out the implementation without many other
changes, like the tests, thereby making this more maintainable.

We are in fact doing this so we can almost immediately switch out the
implementation to use JSON input data instead of HTML.