No need to manipulate raw time figure any more
In the JSON it's a UTC time so we can just parse it directly.

The JSON also has an TimeZoneOffset field that can be used to calculate
the local time but since we want to store it in the database as UTC
anyway we don't need to do any more magic.

We're also removing the tests related to the unneccessary manipulation
too, but leaving the ones that allow the value to be nil.