Replace Capybara scraper with JSON scraper and static sites list
This creates a new static list of the air quality monitoring sites and
their JSON URLs and starts parsing them instead of the HTML pages.

Removes the auto-discovery of sites but I think this is an easy
trade-off for the reduction in complexity of the code. If other sites
are ever added (which is unlikely) then it's straightforward to add
support for them.