Rename database.rb something more sensible
This file only contains the definition of the AqmRecord model so it
makes sense to call it that. Our other classes like this are in lib/ so
I think it's good to group them there, even if this is a database model.

This also has the nice side effect of all the other code not needing to
require the /database/ but the model name instead (which itself then
requires the database connection). I think it's neat, anyway :smile: