Make sure AqmReading handles missing measurements correctly
When the upstream data source has problems like faulty equipment the
web page lists values as "-", which we supported when we were parsing
the HTML. This change ensures these missing values are supported in the
JSON, which represents these values (I think!) more correctly as null.

To do this I've added a quick test using all null values and that
exposed a bunch of problems with my implementation, which I've also now
fixed.