Replace use of dotenv and hardcoded database config with YAML file
Previously the project contained hardcoded configuration for the
database and allowed you to configure a local password using a .env
file.

This didn't allow you to configure most aspects of the database, such as
the database name. While it allowed you to configure the local
development password I'd contend that this isn't the most important
thing.

We now use a YAML file to configure all aspects of the database for your
local test and development environment. This has the advantage of being
familiar (it's how Rails usually works) and allows you to configure all
aspects of the database, including the test environment.